+++ crates/actions/src/actions/sync_workspace.rs
@@ -1,4 +1,5 @@
 use crate::operations::*;
+use crate::plugins::*;
 use crate::utils::should_skip_action;
 use miette::IntoDiagnostic;
 use moon_action::{Action, ActionStatus, Operation};
@@ -21,7 +22,7 @@ pub async fn sync_workspace(
     workspace_graph: Arc<WorkspaceGraph>,
     toolchain_registry: Arc<ToolchainRegistry>,
 ) -> miette::Result<ActionStatus> {
-    let _lock = app_context.cache_engine.create_lock("syncWorkspace")?;
+    let _lock = app_context.cache_engine.create_lock(action.get_prefix())?;
 
     // Connect to the remote service in this action,
     // as it always runs before tasks, and we don't need it
@@ -51,7 +52,7 @@ pub async fn sync_workspace(
 
         operation_futures.push(task::spawn(async move {
             Ok(vec![
-                Operation::sync_operation("system:configSchemas")
+                Operation::sync_operation("config-schemas")?
                     .track_async_with_check(
                         || sync_config_schemas(&app_context, false),
                         |result| result,
@@ -71,7 +72,7 @@ pub async fn sync_workspace(
 
         operation_futures.push(task::spawn(async move {
             Ok(vec![
-                Operation::sync_operation("system:codeowners")
+                Operation::sync_operation("codeowners")?
                     .track_async_with_check(
                         || sync_codeowners(&app_context, &workspace_graph, false),
                         |result| result.is_some(),
@@ -92,7 +93,7 @@ pub async fn sync_workspace(
 
         operation_futures.push(task::spawn(async move {
             Ok(vec![
-                Operation::sync_operation("system:vcsHooks")
+                Operation::sync_operation("vcs-hooks")?
                     .track_async_with_check(|| sync_vcs_hooks(&app_context, false), |result| result)
                     .await?,
             ])
@@ -102,19 +103,20 @@ pub async fn sync_workspace(
     if toolchain_registry.has_plugins() {
         debug!("Syncing operations from toolchains");
 
+        let app_context = Arc::clone(&app_context);
+
         operation_futures.push(task::spawn(async move {
             let mut ops = vec![];
 
             for sync_result in toolchain_registry
-                .sync_workspace_all(|registry, _| SyncWorkspaceInput {
+                .sync_workspace_all(|registry, toolchain| SyncWorkspaceInput {
                     context: registry.create_context(),
+                    toolchain_config: registry
+                        .create_config(&toolchain.id, &app_context.toolchain_config),
                 })
                 .await?
             {
-                ops.push(convert_plugin_sync_operation_with_output(
-                    sync_result.operation,
-                    sync_result.output,
-                ));
+                ops.push(finalize_sync_operation(sync_result)?);
             }
 
             Ok(ops)