+++ legacy/node/tool/src/pnpm_tool.rs
@@ -108,7 +108,12 @@ impl Tool for PnpmTool {
         self.console
             .print_checkpoint(Checkpoint::Setup, format!("installing pnpm {version}"))?;
 
-        if self.tool.setup(version, InstallOptions::default()).await? {
+        if self
+            .tool
+            .setup(version, InstallOptions::default())
+            .await?
+            .is_some()
+        {
             last_versions.insert("pnpm".into(), version.to_owned());
             count += 1;
         }