+++ legacy/rust/tool/src/rust_tool.rs
@@ -156,7 +156,12 @@ impl Tool for RustTool {
                 self.console
                     .print_checkpoint(Checkpoint::Setup, format!("installing rust {version}"))?;
 
-                if self.tool.setup(version, InstallOptions::default()).await? {
+                if self
+                    .tool
+                    .setup(version, InstallOptions::default())
+                    .await?
+                    .is_some()
+                {
                     last_versions.insert("rust".into(), version.to_owned());
                     installed += 1;
                 }